Dynamic Panel Model (dpm) class

Models fit using dpm() return values of this class, which inherits from lavaan-class.

Slots

call_info

A list of metadata about the arguments used.

call

The actual function call.

mod_string

The model formula passed to lavaan.

wide_data

The data provided to the data argument in the function

formula

The Formula::Formula() object provided to dpm(). call.