lavaan-style summary for dpm objects

This is just a quick way to get lavaan's summary instead the more terse summary designed for dpm objects.

lav_summary(x, ...)

Arguments

x

The dpm object

...

Other arguments to the lavaan function.

Value

Returns a lavaan.summary object which contains various model data described in ?lavaan::summary,lavaan-method.

Examples

# Load example data
data("WageData", package = "panelr")
# Convert data to panel_data format for ease of use
wages <- panel_data(WageData, id = id, wave = t)

fit <- dpm(wks ~ pre(lag(union)) + lag(lwage) | ed, data = wages)
lav_summary(fit)
#> lavaan 0.6.17 ended normally after 594 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                       179
#>   Number of equality constraints                    20
#> 
#>   Number of observations                           595
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               110.228
#>   Degrees of freedom                                71
#>   P-value (Chi-square)                           0.002
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   alpha =~                                            
#>     wks_2             1.000                           
#>     wks_3             1.000                           
#>     wks_4             1.000                           
#>     wks_5             1.000                           
#>     wks_6             1.000                           
#>     wks_7             1.000                           
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   wks_2 ~                                             
#>     union_1  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_1  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_1     (p1)    0.187    0.020    9.275    0.000
#>   wks_3 ~                                             
#>     union_2  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_2  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_2     (p1)    0.187    0.020    9.275    0.000
#>   wks_4 ~                                             
#>     union_3  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_3  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_3     (p1)    0.187    0.020    9.275    0.000
#>   wks_5 ~                                             
#>     union_4  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_4  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_4     (p1)    0.187    0.020    9.275    0.000
#>   wks_6 ~                                             
#>     union_5  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_5  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_5     (p1)    0.187    0.020    9.275    0.000
#>   wks_7 ~                                             
#>     union_6  (en1)   -1.191    0.513   -2.321    0.020
#>     lwage_6  (ex1)    0.642    0.484    1.327    0.185
#>     ed        (c1)   -0.112    0.056   -2.011    0.044
#>     wks_6     (p1)    0.187    0.020    9.275    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   alpha ~~                                            
#>     union_1          -0.116    0.109   -1.065    0.287
#>     union_2          -0.106    0.111   -0.953    0.341
#>     union_3          -0.074    0.114   -0.651    0.515
#>     union_4          -0.073    0.116   -0.627    0.530
#>     union_5          -0.094    0.119   -0.786    0.432
#>     union_6          -0.083    0.120   -0.694    0.488
#>     lwage_1          -0.000    0.068   -0.006    0.996
#>     lwage_2           0.001    0.064    0.008    0.993
#>     lwage_3           0.025    0.078    0.324    0.746
#>     lwage_4           0.054    0.076    0.715    0.475
#>     lwage_5           0.029    0.073    0.390    0.697
#>     lwage_6           0.034    0.073    0.466    0.641
#>     wks_1             3.987    0.808    4.938    0.000
#>  .wks_5 ~~                                            
#>     union_6           0.045    0.034    1.294    0.196
#>  .wks_4 ~~                                            
#>     union_5          -0.024    0.035   -0.683    0.495
#>     union_6           0.020    0.040    0.502    0.616
#>  .wks_3 ~~                                            
#>     union_4           0.022    0.033    0.661    0.509
#>     union_5           0.001    0.043    0.019    0.985
#>     union_6          -0.004    0.047   -0.084    0.933
#>  .wks_2 ~~                                            
#>     union_3          -0.100    0.041   -2.444    0.015
#>     union_4          -0.078    0.043   -1.823    0.068
#>     union_5          -0.115    0.052   -2.214    0.027
#>     union_6          -0.067    0.055   -1.216    0.224
#>   union_1 ~~                                          
#>     ed               -0.379    0.057   -6.649    0.000
#>     lwage_1           0.012    0.008    1.507    0.132
#>     lwage_2           0.015    0.007    2.031    0.042
#>     lwage_3          -0.001    0.009   -0.169    0.866
#>     lwage_4          -0.006    0.009   -0.716    0.474
#>     lwage_5          -0.003    0.008   -0.341    0.733
#>     lwage_6          -0.001    0.008   -0.100    0.920
#>     wks_1            -0.401    0.124   -3.227    0.001
#>   ed ~~                                               
#>     union_2          -0.346    0.056   -6.154    0.000
#>   lwage_1 ~~                                          
#>     union_2           0.010    0.008    1.360    0.174
#>   union_2 ~~                                          
#>     lwage_2           0.013    0.007    1.813    0.070
#>     lwage_3          -0.002    0.009   -0.242    0.809
#>     lwage_4          -0.007    0.009   -0.777    0.437
#>     lwage_5          -0.002    0.008   -0.292    0.771
#>     lwage_6           0.002    0.008    0.246    0.806
#>   union_1 ~~                                          
#>     union_2           0.210    0.013   16.512    0.000
#>   wks_1 ~~                                            
#>     union_2          -0.449    0.123   -3.638    0.000
#>   ed ~~                                               
#>     union_3          -0.333    0.057   -5.877    0.000
#>   lwage_1 ~~                                          
#>     union_3           0.006    0.008    0.782    0.434
#>   lwage_2 ~~                                          
#>     union_3           0.008    0.007    1.141    0.254
#>   union_3 ~~                                          
#>     lwage_3          -0.004    0.009   -0.420    0.675
#>     lwage_4          -0.007    0.009   -0.779    0.436
#>     lwage_5          -0.004    0.008   -0.471    0.638
#>     lwage_6           0.004    0.008    0.435    0.664
#>   union_1 ~~                                          
#>     union_3           0.199    0.013   15.911    0.000
#>   union_2 ~~                                          
#>     union_3           0.206    0.013   16.274    0.000
#>   wks_1 ~~                                            
#>     union_3          -0.595    0.126   -4.725    0.000
#>   ed ~~                                               
#>     union_4          -0.384    0.057   -6.694    0.000
#>   lwage_1 ~~                                          
#>     union_4           0.004    0.008    0.455    0.649
#>   lwage_2 ~~                                          
#>     union_4           0.007    0.007    0.915    0.360
#>   lwage_3 ~~                                          
#>     union_4          -0.006    0.009   -0.719    0.472
#>   union_4 ~~                                          
#>     lwage_4          -0.010    0.009   -1.114    0.265
#>     lwage_5          -0.006    0.008   -0.717    0.473
#>     lwage_6           0.001    0.008    0.155    0.877
#>   union_1 ~~                                          
#>     union_4           0.201    0.013   15.983    0.000
#>   union_2 ~~                                          
#>     union_4           0.205    0.013   16.202    0.000
#>   union_3 ~~                                          
#>     union_4           0.211    0.013   16.407    0.000
#>   wks_1 ~~                                            
#>     union_4          -0.552    0.126   -4.389    0.000
#>   ed ~~                                               
#>     union_5          -0.348    0.057   -6.129    0.000
#>   lwage_1 ~~                                          
#>     union_5           0.000    0.008    0.024    0.981
#>   lwage_2 ~~                                          
#>     union_5           0.005    0.007    0.696    0.487
#>   lwage_3 ~~                                          
#>     union_5          -0.008    0.009   -0.872    0.383
#>   lwage_4 ~~                                          
#>     union_5          -0.010    0.009   -1.172    0.241
#>   union_5 ~~                                          
#>     lwage_5          -0.008    0.008   -0.907    0.365
#>     lwage_6          -0.002    0.008   -0.213    0.831
#>   union_1 ~~                                          
#>     union_5           0.190    0.012   15.507    0.000
#>   union_2 ~~                                          
#>     union_5           0.197    0.012   15.884    0.000
#>   union_3 ~~                                          
#>     union_5           0.202    0.013   16.040    0.000
#>   union_4 ~~                                          
#>     union_5           0.210    0.013   16.337    0.000
#>   wks_1 ~~                                            
#>     union_5          -0.586    0.126   -4.664    0.000
#>   ed ~~                                               
#>     union_6          -0.370    0.057   -6.506    0.000
#>   lwage_1 ~~                                          
#>     union_6           0.003    0.008    0.368    0.713
#>   lwage_2 ~~                                          
#>     union_6           0.006    0.007    0.909    0.363
#>   lwage_3 ~~                                          
#>     union_6          -0.008    0.009   -0.856    0.392
#>   lwage_4 ~~                                          
#>     union_6          -0.009    0.009   -1.061    0.289
#>   lwage_5 ~~                                          
#>     union_6          -0.006    0.008   -0.758    0.448
#>   union_6 ~~                                          
#>     lwage_6          -0.000    0.008   -0.022    0.983
#>   union_1 ~~                                          
#>     union_6           0.188    0.012   15.427    0.000
#>   union_2 ~~                                          
#>     union_6           0.194    0.012   15.804    0.000
#>   union_3 ~~                                          
#>     union_6           0.198    0.012   15.866    0.000
#>   union_4 ~~                                          
#>     union_6           0.207    0.013   16.253    0.000
#>   union_5 ~~                                          
#>     union_6           0.214    0.013   16.609    0.000
#>   wks_1 ~~                                            
#>     union_6          -0.455    0.124   -3.667    0.000
#>   lwage_1 ~~                                          
#>     ed                0.426    0.048    8.944    0.000
#>     wks_1             0.316    0.100    3.152    0.002
#>   ed ~~                                               
#>     lwage_2           0.422    0.045    9.399    0.000
#>   lwage_1 ~~                                          
#>     lwage_2           0.132    0.008   16.724    0.000
#>   wks_1 ~~                                            
#>     lwage_2           0.256    0.093    2.745    0.006
#>   ed ~~                                               
#>     lwage_3           0.526    0.055    9.505    0.000
#>   lwage_1 ~~                                          
#>     lwage_3           0.149    0.009   15.898    0.000
#>   lwage_2 ~~                                          
#>     lwage_3           0.140    0.009   15.976    0.000
#>   wks_1 ~~                                            
#>     lwage_3           0.364    0.115    3.156    0.002
#>   ed ~~                                               
#>     lwage_4           0.550    0.055    9.980    0.000
#>   lwage_1 ~~                                          
#>     lwage_4           0.145    0.009   15.760    0.000
#>   lwage_2 ~~                                          
#>     lwage_4           0.138    0.009   15.948    0.000
#>   lwage_3 ~~                                          
#>     lwage_4           0.177    0.011   16.307    0.000
#>   wks_1 ~~                                            
#>     lwage_4           0.359    0.114    3.157    0.002
#>   ed ~~                                               
#>     lwage_5           0.540    0.053   10.148    0.000
#>   lwage_1 ~~                                          
#>     lwage_5           0.140    0.009   15.817    0.000
#>   lwage_2 ~~                                          
#>     lwage_5           0.134    0.008   16.033    0.000
#>   lwage_3 ~~                                          
#>     lwage_5           0.167    0.010   16.149    0.000
#>   lwage_4 ~~                                          
#>     lwage_5           0.169    0.010   16.385    0.000
#>   wks_1 ~~                                            
#>     lwage_5           0.308    0.109    2.819    0.005
#>   ed ~~                                               
#>     lwage_6           0.524    0.053    9.893    0.000
#>   lwage_1 ~~                                          
#>     lwage_6           0.134    0.009   15.425    0.000
#>   lwage_2 ~~                                          
#>     lwage_6           0.128    0.008   15.631    0.000
#>   lwage_3 ~~                                          
#>     lwage_6           0.162    0.010   15.853    0.000
#>   lwage_4 ~~                                          
#>     lwage_6           0.165    0.010   16.159    0.000
#>   lwage_5 ~~                                          
#>     lwage_6           0.166    0.010   16.584    0.000
#>   wks_1 ~~                                            
#>     lwage_6           0.270    0.109    2.475    0.013
#>   ed ~~                                               
#>     wks_1             0.067    0.714    0.094    0.925
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .wks_2            36.140    2.893   12.492    0.000
#>    .wks_3            35.953    2.936   12.247    0.000
#>    .wks_4            36.037    2.995   12.033    0.000
#>    .wks_5            35.719    3.041   11.745    0.000
#>    .wks_6            35.464    3.080   11.514    0.000
#>    .wks_7            35.177    3.114   11.297    0.000
#>     union_1           0.361    0.020   18.348    0.000
#>     lwage_1           6.375    0.016  400.689    0.000
#>     ed               12.845    0.114  112.400    0.000
#>     wks_1            46.281    0.256  180.680    0.000
#>     union_2           0.348    0.020   17.817    0.000
#>     lwage_2           6.465    0.015  435.167    0.000
#>     union_3           0.370    0.020   18.697    0.000
#>     lwage_3           6.597    0.018  360.533    0.000
#>     union_4           0.373    0.020   18.820    0.000
#>     lwage_4           6.696    0.018  370.896    0.000
#>     union_5           0.366    0.020   18.550    0.000
#>     lwage_5           6.786    0.017  390.740    0.000
#>     union_6           0.363    0.020   18.459    0.000
#>     lwage_6           6.864    0.017  395.200    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .wks_2            18.724    1.200   15.601    0.000
#>    .wks_3            14.630    0.966   15.144    0.000
#>    .wks_4            11.995    0.814   14.744    0.000
#>    .wks_5            16.314    1.057   15.430    0.000
#>    .wks_6            16.400    1.060   15.478    0.000
#>    .wks_7            17.846    1.137   15.702    0.000
#>     union_1           0.231    0.013   17.248    0.000
#>     lwage_1           0.151    0.009   17.248    0.000
#>     ed                7.771    0.451   17.248    0.000
#>     wks_1            39.039    2.263   17.248    0.000
#>     union_2           0.227    0.013   17.248    0.000
#>     lwage_2           0.131    0.008   17.248    0.000
#>     union_3           0.233    0.013   17.276    0.000
#>     lwage_3           0.199    0.012   17.248    0.000
#>     union_4           0.234    0.014   17.268    0.000
#>     lwage_4           0.194    0.011   17.248    0.000
#>     union_5           0.232    0.013   17.281    0.000
#>     lwage_5           0.179    0.010   17.248    0.000
#>     union_6           0.230    0.013   17.271    0.000
#>     lwage_6           0.179    0.010   17.248    0.000
#>     alpha             4.422    0.553    7.999    0.000
#>