Summarize dpm objects

The summary method is designed to offer similar arguments to lavaan's summary, but with shorter and more domain-specific output.

# S4 method for dpm
summary(
  object,
  standardized = FALSE,
  ci = FALSE,
  se = TRUE,
  zstat = TRUE,
  pvalue = TRUE,
  ci.level = 0.95,
  boot.ci.type = c("perc", "norm", "basic", "bca.simple"),
  digits = getOption("dpm-digits", 3),
  ...
)

Arguments

object

A dpm object.

standardized

Use lavaan's method for standardizing coefficients? Default is FALSE.

ci

Show confidence intervals? Default is FALSE.

se

Show standard errors? Default is TRUE.

zstat

Show the z statistic? Default is TRUE.

pvalue

Show p values? Default is TRUE.

ci.level

How wide should the confidence intervals be? Ignored if ci is FALSE. Default is .95.

boot.ci.type

If the model was fit with bootstrapped standard errors and ci is TRUE, which method should be used for finding the intervals? Default is "perc".

digits

How many digits should be printed in the model summary? Default is 3. You can set a default by setting the option "dpm-digits".

...

Ignored.

Value

Returns a summary.dpm object, which is a list with three elements:

The dpm object.

A data frame containing coefficient estimates, standard errors, p values, and so on.)

A numeric vector containing model fit information.

The primary function of the object is to be printed to the console.